move to folder and add constants file
