+++ frontend/src/components/Snackbar/SnackbarSeverityConstants.js
@@ -1,0 +1,4 @@
+export const SUCCESS = "success"
+export const ERROR = "error"
+export const INFO = "info"
+export const WARNING = "warning"
